refactor(plugin-space): move makeCreateObjectEntryForDatabaseType to util
