+++ packages/plugins/plugin-space/src/capabilities/index.ts
@@ -6,7 +6,7 @@ import { Capability } from '@dxos/app-framework';
 import { OperationHandlerSet } from '@dxos/compute';
 
 export * from './app-graph-builder';
-export { makeCreateObjectEntryForDatabaseType } from './make-create-entry';
+export { makeCreateObjectEntryForDatabaseType } from '../util';
 
 export const AppGraphSerializer = Capability.lazy('AppGraphSerializer', () => import('./app-graph-serializer'));
 export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
