+++ packages/plugins/plugin-space/src/util.ts
@@ -2,10 +2,16 @@
 // Copyright 2023 DXOS.org
 //
 
+import * as Effect from 'effect/Effect';
+
 import { type Space, SpaceState } from '@dxos/client/echo';
+import { Operation } from '@dxos/compute';
+import { Obj, Type } from '@dxos/echo';
 import { type Label } from '@dxos/ui-types/translations';
 
 import { meta } from '#meta';
+import { SpaceOperation } from '#operations';
+import { type SpaceCapabilities } from '#types';
 
 //
 // Constants
@@ -20,6 +26,24 @@ export { SHARED } from './types';
 //
 // Helpers
 //
+
+/**
+ * Creates a synthetic CreateObjectEntry for a database-persisted object schema that has no dedicated plugin capability.
+ * Sets inputSchema so the create dialog presents a form for the type's fields.
+ */
+export const makeCreateObjectEntryForDatabaseType = (type: Type.AnyObj): SpaceCapabilities.CreateObjectEntry => ({
+  id: Type.getTypename(type),
+  inputSchema: Type.getSchema(type),
+  createObject: (props, options) =>
+    Effect.gen(function* () {
+      const object = Obj.make(type, props);
+      return yield* Operation.invoke(SpaceOperation.AddObject, {
+        object,
+        target: options.target,
+        hidden: true,
+      });
+    }),
+});
 
 /** Returns the display label for a space (name, namesCache entry, or fallback). */
 export const getSpaceDisplayName = (
